requirements.txt
aiogram==3.20.0
python-dotenv==1.1.0
asyncpg==0.30.0